Amounts (#95)
* Fixing bity amount too low errors

* split payment state/reducer

* max output amount to 1000

* walkthrough in about
